Salad-builder app I am to work with for chapter 6
After adding the code in the salad-builder folder to our own repository we were to run "npm install" for adding dependencies and then "npm start" to run the application. But it gave me issues due to the dependencies being unsupported and outdated because the book was made a couple years back and things progress so rapidly. (The instructions I gave are given on page 104 in the book).